Fix crash when 2nd call has been accepted on OSX. (Delete procee handle twice.) Fix thread id bug in Logger.

+++ src/ext/gwutils/Logger.h
@@ -53,7 +53,7 @@ private:
 	void Unlock();
 	void FillLog(const int nLevel, const char* szTAG, const char* szMessage);
 	void FillCurrentDateTimeString(char* szBuf, const int nBufSize);
-	unsigned int GetCurrentThreadID() const;
+	uint64_t GetCurrentThreadID() const;
 
 	void PrintToConsole(const int nLevel, const char* szTAG);
 	void PrintToFile();
